+++ watchdog_server/src/main/java/com/siemt3/watchdog_server/cep/listener/port/PortDistVerticalScanListener.java
@@ -23,11 +23,11 @@ import static com.siemt3.watchdog_server.GlobalVariables.DEBUG_FLAG;
 public class PortDistVerticalScanListener implements UpdateListener {
 	@Override
 	public void update(EventBean[] newEvents, EventBean[] oldEvents, EPStatement statement, EPRuntime runtime) {
-		Long arrivalTime = (long) newEvents[0].get("arrivalTime");
+		long arrivalTime = (long) newEvents[0].get("arrivalTime");
 		String dstIp = (String) newEvents[0].get("dstIp");
-		Integer maxPort = (Integer) newEvents[0].get("maxPort");
-		Integer minPort = (Integer) newEvents[0].get("minPort");
-		Integer srcIpCount = (Integer) newEvents[0].get("srcIpCount");
+		Integer maxPort = (int) newEvents[0].get("maxPort");
+		Integer minPort = (int) newEvents[0].get("minPort");
+		Integer srcIpCount = (int) (long) newEvents[0].get("srcIpCount");
 
 		PortDistVerticalScanEvent portDistVertical = new PortDistVerticalScanEvent(arrivalTime, dstIp, maxPort, minPort,
 				srcIpCount);
